% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.normalize}
\alias{.normalize}
\title{Normalization method for affinity matrices}
\usage{
.normalize(X)
}
\arguments{
\item{X}{matrix. Similarity matrix to normalize.}
}
\value{
Normalized matrix.
}
\description{
Normalization method used to compute a "global" similarity
matrix in SNF algorithm, capturing the overall relationships between patients.
It corresponds to the formula:

\loadmathjax
\mjsdeqn{P(i,j) = 
\begin{cases}
\frac{W(i,j)}{2 \sum_{k \neq i} W(i,k)} & \text{, if $j \neq i$}\\

1/2 & , \text{if $j = i$}
\end{cases}}

NOTE: This function is taken from SNFtool 
(https://cran.r-project.org/web/packages/SNFtool/index.html), but it was
defined inside the main function SNFtool::SNF. Here it is used as internal
function to make the code more readable.
}
