% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missSNF.R
\name{get.miss.pts}
\alias{get.miss.pts}
\title{Get list of missing patients for each matrix}
\usage{
get.miss.pts(Mall, perc.na = 0.2, miss.symbol = NULL)
}
\arguments{
\item{Mall}{list of named matrices/dataframes (samples x features).}

\item{perc.na}{percentage of NAs above which patient is considered missing.}

\item{miss.symbol}{if not NULL, the provided symbol in matrices is converted
to NA.}
}
\value{
list of vectors. Each vector contains the indices of missing samples
        in a specific matrix.
}
\description{
This function takes in input a list of matrices Mall and returns
for each matrix the list of patients that are missing (considering the union
of patients in all matrices). A patient is considered "missing" if (I) it is
present in the matrix but has |NA| > perc.na, (II) a patient is not present
in one matrix but is present in at least one of the others.
}
