% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missSNF.R
\name{impute_miss}
\alias{impute_miss}
\title{Impute patients with few NAs}
\usage{
impute_miss(data, perc.na, method, verbose = TRUE)
}
\arguments{
\item{data}{matrix. Matrix (samples x features).}

\item{perc.na}{numeric. Percentage of missing features.}

\item{method}{string. An imputation method (possible options mean, median).}

\item{verbose}{boolean. Want to print messages? (def. TRUE)}
}
\value{
Imputed data matrix.
}
\description{
This function imputes missing samples using some specified
method (e.g. mean, median) only for patients/samples having a percentage
of NAs lower or equal to perc.na. In other words, not all missing data are
imputed but only the ones for which a patient has enough data to not be
considered completely missing.
}
