% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missSNF.R
\name{miss.snf}
\alias{miss.snf}
\title{miss-SNF: integration of patients' networks considering missing cases}
\usage{
miss.snf(
  Mall,
  sims,
  mode = "reconstruct",
  perc.na = 0.2,
  miss.symbols = NULL,
  K = 20,
  t = 20,
  impute = "median",
  d = 1,
  random.walk = "none",
  p = 3,
  ...
)
}
\arguments{
\item{Mall}{list of named matrices/dataframes (samples x features).}

\item{sims}{vector of strings. It is a vector containing the names of the
similarity measures to apply to the matrices in Mall.
"scaled.exp.euclidean" is the scaled exponential euclidean distance;
"scaled.exp.chi2" is the scaled exponential chi-square distance}

\item{mode}{string. If you want to partially reconstruct missing data use
"reconstruct", otherwise ignore them during integration using
"ignore".}

\item{perc.na}{percentage of NAs above which patient is considered missing.}

\item{miss.symbols}{vector of strings. If not NULL, the provided symbols
in matrices are converted to NA.}

\item{K}{Number of neighbors in K-nearest neighbors part of the algorithm.}

\item{t}{Number of iterations for the diffusion process.}

\item{impute}{string. Kind of imputation method to apply in case of samples
with few missing values. Options are: NULL, "mean", "median".
NOTE: if impute=NULL, then perc.na has to be 0.}

\item{d}{numeric. Set the diagonal of the matrix to "d" if
mode="reconstruct" (def d=1).}

\item{random.walk}{string. Use 1-step Random Walk to compute the local
similarity matrix S and/or p-step Random Walk to compute the global
similarity matrix P. random.walk=c("global", "local", "both", "none") and
defaults is random.walk="none".}

\item{p}{numeric. Number of steps for the p-step RW.}

\item{...}{additional arguments for similarity measures.}
}
\value{
A list with two elements:
\itemize{
\item W : integrated similarity matrix. Note that the order of the patients
is different from the order of the original matrices.
\item removed.pts : vector with names of removed patients (i.e. patients
present only in one matrix of Mall and having too much NAs).
}
}
\description{
Extension of the algorithm Similarity Network Fusion
(\url{https://doi.org/10.1038/nmeth.2810}) able to handle the absence (complete or
nearly complete) of a specific data source for a given patient by:
\loadmathjax
\itemize{
\item "reconstruct strategy" that can partially reconstruct missing
  data by using information from different sources
\item "ignore strategy" which simply ignores missing data during the integration
  process
}
}
\examples{

# Create list of input matrices
set.seed(123);
M1 <- matrix(runif(50, min = 0, max = 1), nrow = 5); # 5 samples
rownames(M1) <- paste0("ID_", 1:nrow(M1));
M1[1, 1:4] <- c(NA, NA, NA, NA);
M1[4, ] <- rep(NA, ncol(M1));

M2 <- matrix(runif(80, min = 0, max = 1), nrow = 8); # 8 samples
rownames(M2) <- paste0("ID_", 3:10);
M2[2, ] <- rep(NA, ncol(M2));
M2[4, ] <- rep(NA, ncol(M2));
M2[8, 1] <- NA;

M3 <- matrix(runif(80, min = 0, max = 1), nrow = 8); # 8 samples
rownames(M3) <- c(paste0("ID_", 1:5), paste0("ID_", 11:13));
M3[4, 1] <- NA;
M3[7, 1:8] <- rep(NA, 8);

Mall <- list("M1"=M1, "M2"=M2, "M3"=M3);

# Call miss.snf using "reconstruct strategy"
W.r <- miss.snf(Mall, sims=rep("scaled.exp.euclidean", 3), mode="reconstruct",
               K=3, kk=2);

# Call miss.snf using "ignore strategy"
W.i <- miss.snf(Mall, sims=rep("scaled.exp.euclidean", 3), mode="ignore",
               K=3, kk=2);
}
