% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missSNF.R
\name{scaled.exp.chi2}
\alias{scaled.exp.chi2}
\title{Scaled exponential chi-square distance}
\usage{
scaled.exp.chi2(M, kk = 20, sigma = 0.5)
}
\arguments{
\item{M}{matrix (samples x features)}

\item{kk}{integer. Number of nearest neighbours to sparsify similarity.}

\item{sigma}{numeric. Variance for local model.}
}
\value{
similarity matrix computed using scaled exponential chi-square
        distance.
}
\description{
it is a wrapper function to directly compute the scaled
exponential chi-square distance as implemented in SNFtool library.
}
