% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missSNF.R
\name{scaled.exp.chi2}
\alias{scaled.exp.chi2}
\title{Scaled exponential chi-square distance}
\usage{
scaled.exp.chi2(M, kk = 20, sigma = 0.5)
}
\arguments{
\item{M}{matrix/dataframe (samples x features).}

\item{kk}{integer. Number of nearest neighbours to sparsify similarity.}

\item{sigma}{numeric. Variance for local model.}
}
\value{
similarity matrix computed using scaled exponential chi-square
        distance.
}
\description{
It is a wrapper function to directly compute the scaled
exponential chi-square distance as implemented in SNFtool library.
}
\examples{

# Create a matrix
set.seed(123);
M1 <- rbind(c(1,2,1,1,1,2,2), c(2,2,2,1,1,1,1), c(1,2,1,2,1,2,1),
           c(2,2,2,1,2,1,2));

rownames(M1) <- paste0("ID_", 1:nrow(M1));

# Compute similarity matrix
sim <- scaled.exp.chi2(M1, kk=3);
}
